moved tutorcard into it's own folder and added index.js to complete the export of it.  Deleted several images that we're not using, deleted the Homepage2 page, revised the props functions passing into the StudentCard (still needs work)
