--- client/Images/Working-it.mp4
@@ -1,2 +1,0 @@
-[InternetShortcut]
-URL=https://github.com/DizDonut/tutorBook/blob/master/Images%20/Working-it.mp4
