--- client/Images/images2.jpg
@@ -1,2 +1,0 @@
-[InternetShortcut]
-URL=https://github.com/DizDonut/tutorBook/blob/master/Images%20/images2.jpg
