--- client/Images/math_tutoring_WebO.jpg
@@ -1,2 +1,0 @@
-[InternetShortcut]
-URL=https://github.com/DizDonut/tutorBook/blob/master/Images%20/math_tutoring_WebO.jpg
