--- client/Images/ressources.jpg
@@ -1,2 +1,0 @@
-[InternetShortcut]
-URL=https://github.com/DizDonut/tutorBook/blob/master/Images%20/ressources.jpg
