+++ client/src/App.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import Homepage from "./pages/Homepage";
-import Homepage2 from "./pages/Homepage2";
 import Tutors from "./pages/Tutors";
 import Student from "./pages/Student";
