+++ client/src/components/StudentCard/StudentCard.js
@@ -3,15 +3,17 @@ import { Card, CardTitle, Col, Container, Row } from "react-materialize";
 import Nav from "../../components/Navbar";
 
 const StudentCard = props => {
-  <Row>
-    <Col s={3}>
-      <Card className="small">
-          <CardTitle reveal image={props} waves="light" />
-            {props.children}
-      </Card>
-    </Col>
-  </Row>
-
-} //end StudentCard const
+  return(
+    <Row>
+      <Col s={3}>
+        <Card className="small">
+            <CardTitle reveal image={props.reveal} waves="light" />
+              {props.header}
+              {props.title}
+        </Card>
+      </Col>
+    </Row>
+  );
+}; //end StudentCard const
 
 export default StudentCard;
