+++ client/src/components/Tutor/index.js
@@ -1,0 +1,1 @@
+export { default } from "./TutorCard.js";
