+++ client/src/pages/Homepage/Homepage.js
@@ -14,16 +14,16 @@ class Homepage extends Component {
       <Nav />
 
      <Container>
-  
+
        <Card className="large red lighten-5"
-        header={<CardTitle image="../../../Images/homepage.jpg" ><h2>Welcome To Bao Bao Book!</h2> </CardTitle>}
+        header={<CardTitle image={"../../../Images/homepage.jpg"}><h2>Welcome To Bao Bao Book!</h2> </CardTitle>}
         >
         <p>Lorem ipsum dolor sit amet, integer quis vulputate, vel purus nullam consectetuer, nullam interdum dapibus, amet sed sagittis tincidunt libero eros amet, quam mattis. Sem orci arcu ipsum vitae consectetuer lorem. Et arcu vitae habitant, leo ut quam condimentum mollis, augue enim ultrices,Lorem ipsum dolor sit amet, integer quis vulputate, vel purus nullam consectetuer, nullam interdum dapibus, amet sed sagittis tincidunt libero eros amet, quam mattis. Sem orci arcu ipsum vitae consectetuer lorem. Et arcu vitae habitant, leo ut quam condimentum mollis, augue enim.Lorem ipsum dolor sit amet, integer quis vulputate, vel purus nullam consectetuer, nullam interdum dapibus, amet sed sagittis tincidunt libero eros amet, quam mattis. Sem orci arcu ipsum vitae consectetuer lorem. Et arcu vitae habitant, leo ut quam condimentum mollis, augue enim ultrices,Lorem ipsum dolor sit amet, integer quis vulputate, vel purus nullam consectetuer, nullam interdum dapibus, amet sed sagittis tincidunt libero eros amet, quam mattis. Sem orci arcu ipsum vitae consectetuer lorem. Et arcu vitae habitant </p>
        </Card>
 
         <Row>
           <Col s={12}>
-           
+
             {/*<Link to={"/tutors/"}>To get to tutors page</Link> {/* will need to add + tutors._id*//*}*/}
             <div className="row center">
               <Row>
