--- client/src/pages/Homepage2/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Homepage2.js";
